+++ assets/UsagePage-ee5928da.js
@@ -1,0 +1,12 @@
+var a=Object.defineProperty;var r=(e,o)=>a(e,"name",{value:o,configurable:!0});import{L as i}from"./Link-af594e99.js";import{d as t,T as n,aF as s}from"./index-d23866de.js";import{P as m}from"./PaperCode-cf567731.js";import"./copyText-288aae3f.js";import"./_commonjsHelpers-0119a68f.js";const l=`// @refresh
+import { Button } from "@suid/material";
+import { render } from "solid-js/web";
+
+function App() {
+  return <Button variant="contained">Hello World</Button>;
+}
+
+// eslint-disable-next-line @typescript-eslint/no-non-null-assertion
+render(() => <App />, document.getElementById("root")!);
+`;function f(){return[t(n,{component:"h1",variant:"h4",sx:{mt:1},children:"Usage"}),t(n,{variant:"body1",sx:{mt:2},get children(){return["For a completed usage guide, please go to"," ",t(i,{href:"https://mui.com",target:"_blank",children:"MUI"}),"."]}}),t(n,{component:"h2",variant:"h5",sx:{mt:2,mb:1},children:"Quick usage"}),t(m,{language:"tsx",get value(){return l.split(/\r?\n/).filter(e=>!e.trim().startsWith("// eslint-disable")).join(`
+`)}}),t(s,{sx:{mt:2}})]}r(f,"UsagePage");export{f as default};