+++ assets/createUniqueId-5e414aed.js
@@ -1,0 +1,1 @@
+var t=Object.defineProperty;var r=(e,a)=>t(e,"name",{value:a,configurable:!0});import{O as c,aK as i}from"./index-d23866de.js";function u(e){return c(()=>e?.()??`mui-${i()}`)}r(u,"createUniqueId");export{u as c};